u: config 초기화 위치 수정
