+++ app/engine/entryPoints.ts
@@ -5,8 +5,19 @@ import {
   SimulationDifficulty,
 } from "./core/simulation";
 import { useMemoryState } from "~/store/memory";
-import { config, initializeRouter } from "./settings";
+import { initializeRouter } from "./settings";
+import { term } from ".";
 
+// 전역 시뮬레이션 엔진 인스턴스
+export let config = {
+  requestCount: 1000,
+  difficulty: SimulationDifficulty.Normal,
+  topologyType: TopologyType.Star,
+  nodeCount: 5,
+  coresPerNode: 2,
+  tickInterval: new term.MilliSecond(100),
+  timeScale: 1.0,
+};
 let simulationEngine = SimulationEngine.create().configure(config);
 initializeRouter();
